COMP: Fix sign-compare warning in vtkSlicerSegmentationsModuleLogic
This commit fixes the following warning:

/path/to/Slicer/Modules/Loadable/Segmentations/Logic/vtkSlicerSegmentationsModuleLogic.cxx:1865:69: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   while (typesInFirstCategory.empty() && firstNonEmptyCategoryIndex < categories.size());
                                                                     ^

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=26376
git-svn-id: http://svn.slicer.org/Slicer4/trunk@26376 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/Loadable/Segmentations/Logic/vtkSlicerSegmentationsModuleLogic.cxx
@@ -1862,7 +1862,7 @@ bool vtkSlicerSegmentationsModuleLogic::SetTerminologyToSegmentationFromLabelmap
     {
     this->TerminologiesLogic->GetTypesInTerminologyCategory(terminologyContextName, categories[++firstNonEmptyCategoryIndex], typesInFirstCategory);
     }
-  while (typesInFirstCategory.empty() && firstNonEmptyCategoryIndex < categories.size());
+  while (typesInFirstCategory.empty() && firstNonEmptyCategoryIndex < static_cast<int>(categories.size()));
   if (typesInFirstCategory.empty())
     {
     vtkErrorMacro("SetTerminologyToSegmentationFromLabelmapNode: All categories in terminology context " << terminologyContextName << " are empty");
